Extract getTemplate from invokeFunction for reuse
